requirements.txt
aiohttp>=3.11.18
certifi>=2025.4.26
colorama>=0.4.6
discord.py>=2.5.2
requests>=2.32.3
setuptools>=80.3.1
pyzipper>=0.3.6
pytz>=2025.2
aiohttp-socks>=0.10.1
python-dotenv>=1.1.0
onnxruntime>=1.18.1
matplotlib>=3.10.3
arabic-reshaper>=3.0.0
python-bidi>=0.6.7
pillow>=10.4.0
numpy>=1.26.4
nextcord